+++ src/pages/BoardList.jsx
@@ -4,7 +4,7 @@ import { itemsState } from '../utils/state';
 import CreateBoard from '../components/CreateBoard';
 import Board from '../components/Board';
 import Footer from "../components/Footer"
-import { addItem, deleteItem, checkItem, holdItem, filterItems } from '../utils/helper';
+import { addItem, deleteItem, checkItem, holdItem, filterItems } from '../utils/useItemUtils';
 import '../styles/BoardList.css';
 import '../styles/utils.css';
